Patricle trail additions and implementation of Kyle's point system
Added a particle trail effect to all asteroids:
A particle path clsss was used in conjuntion with the built in unity particle system. The purpose is to create a trailing effect on all asteroids.
When gameObjects containing this script are instantiated, a particle system is spawned on the same location with the same directional movement as that gameObject.
The colliders are also identical to each respective asteroid. When an asteroid is hit, so is the ParticleSystem's collider. When struck the object remains on its
tragectory, however it will stop spawning particles. This allows particles to naturally despawn instead of abruptly disapearing all at once.

Kyles point system added, each asteroid destroyed accumulates 10 points.