Bug fixes
Fixed bug where asteroids would disappear occasionally leaving just a trail.

made mounds a child of camera

Added a deletion box as a child to camera. this box deletes anything that falls outside of view of camer (for some reason the onbecomeinvisible function only applies to the left and right side of the camera.